XW-4232 | preparing fo ab test
