+++ app/components/article-featured-video.js
@@ -32,9 +32,11 @@ export default Component.extend({
 	// reinitialize it as component itself is not destroyed. Could be done with didUpdateAttrs
 	// hook, however it is fired twice with new attributes.
 	videoIdObserver: on('didInsertElement', observer('model.embed.jsParams.videoId', function () {
-		if (!inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'RENDER_BEFORE_EMBER')) {
+		if (!inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'RENDER_ORDER')) {
 			this.destroyVideoPlayer();
 			this.initVideoPlayer();
+		} else {
+			window.updateFeaturedVideoPosition();
 		}
 	})),
