+++ app/modules/video-players/jwplayer-video-ads.js
@@ -1,4 +1,5 @@
 import Ads from '../ads';
+import {Promise} from 'rsvp';
 
 export default class JWPlayerVideoAds {
 	constructor(params) {
